finalizing error handling + documentation
